Feat/fe#389: 가이드 프로필·소개 화면 통합 및 intro 경로 리다이렉트
Made-with: Cursor
